Fixed download_phoenix.
The code is working (tested on Mac OS X), commented, shorter and more readable (I hope).
Also I tried to use standard libraries wherever possible instead of importing external ones.
Now it doesn't depend on changes in URLs (that was why it was broken in the first place - the URLs changed) except the base URL for data versions http://phoenixdata.org/data/<version>/. If it breaks again, it can be easily fixed in a single place (get_links function).